refactor(test): move error tests to the error module
